plugin framework: copy_command refactored as command
